fix interface hot reloading
