+++ .vscode/settings.json
@@ -5,6 +5,7 @@
 		"consts",
 		"countup",
 		"creationdate",
+		"fontsource",
 		"ipfs",
 		"Keepsafe",
 		"pathctx",
@@ -13,6 +14,7 @@
 		"quicktime",
 		"repr",
 		"Roadmap",
+		"subpackage",
 		"svgr",
 		"tailwindcss",
 		"titlebar",
