+++ apps/desktop/src/index.tsx
@@ -65,7 +65,6 @@ function App() {
 
 	return (
 		<SpacedriveInterface
-			useMemoryRouter
 			transport={new Transport()}
 			platform={platform}
 			convertFileSrc={function (url: string): string {
