+++ apps/web/src/App.tsx
@@ -72,7 +72,6 @@ function App() {
 			{/* <header className="App-header"></header> */}
 			<SpacedriveInterface
 				demoMode
-				useMemoryRouter={true}
 				transport={new Transport()}
 				platform={'browser'}
 				convertFileSrc={function (url: string): string {
