+++ packages/interface/src/screens/Debug.tsx
@@ -6,41 +6,32 @@ import { AppPropsContext } from '../App';
 import CodeBlock from '../components/primitive/Codeblock';
 
 export const DebugScreen: React.FC<{}> = (props) => {
-  const appPropsContext = useContext(AppPropsContext);
-  const { data: client } = useBridgeQuery('NodeGetState');
-  const { data: jobs } = useBridgeQuery('JobGetRunning');
-  const { data: jobHistory } = useBridgeQuery('JobGetHistory');
-  // const { mutate: purgeDB } = useBridgeCommand('PurgeDatabase', {
-  //   onMutate: () => {
-  //     alert('Database purged');
-  //   }
-  // });
-  const { mutate: identifyFiles } = useBridgeCommand('IdentifyUniqueFiles');
-  return (
-    <div className="flex flex-col w-full h-screen p-5 custom-scroll page-scroll">
-      <div className="flex flex-col space-y-5 pb-7">
-        <h1 className="text-lg font-bold ">Developer Debugger</h1>
-        <div className="flex flex-row pb-4 space-x-2">
-          <Button
-            className="w-40"
-            variant="gray"
-            size="sm"
-            onClick={() => {
-              if (client && appPropsContext?.onOpen) {
-                appPropsContext.onOpen(client.data_path);
-              }
-            }}
-          >
-            Open data folder
-          </Button>
-
+	const appPropsContext = useContext(AppPropsContext);
+	const { data: client } = useBridgeQuery('NodeGetState');
+	const { data: jobs } = useBridgeQuery('JobGetRunning');
+	const { data: jobHistory } = useBridgeQuery('JobGetHistory');
+	// const { mutate: purgeDB } = useBridgeCommand('PurgeDatabase', {
+	//   onMutate: () => {
+	//     alert('Database purged');
+	//   }
+	// });
+	const { mutate: identifyFiles } = useBridgeCommand('IdentifyUniqueFiles');
+	return (
+		<div className="flex flex-col w-full h-screen p-5 custom-scroll page-scroll">
+			<div className="flex flex-col space-y-5 pb-7">
+				<h1 className="text-lg font-bold ">Developer Debugger</h1>
+				<div className="flex flex-row pb-4 space-x-2">
 					<Button
 						className="w-40"
 						variant="gray"
 						size="sm"
-						onClick={() => identifyFiles(undefined)}
+						onClick={() => {
+							if (client && appPropsContext?.onOpen) {
+								appPropsContext.onOpen(client.data_path);
+							}
+						}}
 					>
-						Identify unique files
+						Open data folder
 					</Button>
 				</div>
 				<h1 className="text-sm font-bold ">Running Jobs</h1>
