+++ packages/interface/src/screens/settings/GeneralSettings.tsx
@@ -15,32 +15,6 @@ export default function GeneralSettings() {
 				title="General Settings"
 				description="Basic settings related to this client."
 			/>
-
-			{/* <InputContainer
-        title="Test scan directory"
-        description="This will create a job to scan the directory you specify to the database."
-      >
-        <div className="flex flex-row">
-          <Input
-            value={tempWatchDir}
-            className="flex-grow"
-            onChange={(e) => setTempWatchDir(e.target.value)}
-            placeholder="/users/jamie/Desktop"
-          />
-          <Button
-            className="ml-2"
-            variant="primary"
-            onClick={() =>
-              createLocation({
-                path: tempWatchDir
-              })
-            }
-          >
-            Scan Now
-          </Button>
-        </div>
-      </InputContainer> */}
-
 			<InputContainer title="Volumes" description="A list of volumes running on this device.">
 				<div className="flex flex-row space-x-2">
 					<div className="flex flex-grow">
